app.py
import streamlit as st 
import openai
import docx
import pypdf

openai.api_key = "your_openai_api_key"

def main():
    st.title("AI-Powered Job Application Assistant")
    
    job_posting = st.text_area("Enter the job posting here")
    resume_file = st.file_uploader("Upload your resume (PDF)", type="pdf")
    
    if st.button("Generate Application"):
        if job_posting and resume_file:
            # Extract text from PDF resume
            resume_text = extract_text_from_pdf(resume_file)
            
            # Generate cover letter using ChatGPT
            cover_letter = generate_cover_letter(job_posting, resume_text)
            
            # Generate tailored resume using ChatGPT 
            tailored_resume = generate_resume(job_posting, resume_text)
            
            # Display preview
            st.markdown("## Cover Letter Preview")
            st.write(cover_letter)
            st.markdown("## Resume Preview") 
            st.write(tailored_resume)
            
            # Export options
            export_cover_letter_pdf(cover_letter)
            export_cover_letter_word(cover_letter)
            export_resume_pdf(tailored_resume)
            export_resume_word(tailored_resume)
        else:
            st.warning("Please provide both the job posting and your resume.")

def extract_text_from_pdf(file):
    # TODO: Implement PDF text extraction using PyPDF
    pass

def generate_cover_letter(job_posting, resume_text):
    # TODO: Implement cover letter generation using ChatGPT API
    pass

def generate_resume(job_posting, resume_text):  
    # TODO: Implement resume generation using ChatGPT API
    pass

def export_cover_letter_pdf(cover_letter):
    # TODO: Implement cover letter PDF export
    pass

def export_cover_letter_word(cover_letter):
    # TODO: Implement cover letter Word export using python-docx
    pass

def export_resume_pdf(resume):
    # TODO: Implement resume PDF export 
    pass
    
def export_resume_word(resume):
    # TODO: Implement resume Word export using python-docx
    pass
        
if __name__ == '__main__':
    main() 