requirements.txt
streamlit==1.32.0
openai==1.12.0
python-docx==1.1.0
pypdf==4.0.1
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.0
markdown==3.5.2
jinja2==3.1.3
pyyaml==6.0.1 